+++ betfairlightweight/resources/bettingresources.py
@@ -580,15 +580,3 @@ class ReplaceOrders(BaseResource):
         sub_resources = {
             'instructionReports': ReplaceOrderInstructionReports
         }
-
-
-class RaceDetails(BaseResource):
-    class Meta(BaseResource.Meta):
-        identifier = 'race_details'
-        attributes = {
-            'responseCode': 'response_code',
-            'raceId': 'race_id',
-            'meetingId': 'meeting_id',
-            'raceStatus': 'race_status',
-            'lastUpdated': 'last_updated'
-        }
